[release/3.5] use triton-lang/llvm-project@triton-3.5
Use a forked branch of llvm-project for the release.

This branch is the same as the original commit hash (064f02dac0c81c19350a74415b3245f42fed09dc) but with https://github.com/llvm/llvm-project/pull/146694 reverted. 146694 results in a divide-by-zero failure (we plan to debug and report an issue or submit a fix, but after the release window) - see https://github.com/pytorch/pytorch/issues/161371 for details. We're also expecting a cherry-picked PR from NVIDIA for https://github.com/llvm/llvm-project/issues/156096 that will require a custom LLVM hash.